using jwt extended

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -16,13 +16,14 @@ class Item(Resource):
                         help="Every item needs a store id."
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -38,13 +39,19 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
 
         return {'message': 'Item deleted'}
 
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
 
@@ -61,5 +68,6 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required
     def get(self):
         return {'items': [x.json() for x in ItemModel.find_all()]}
